Refactored/Improved export/import
